+++ webApp/src/webMain/kotlin/co/edu/unicauca/dopaminah/main.kt
@@ -2,10 +2,13 @@ package co.edu.unicauca.dopaminah
 
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
+import co.edu.unicauca.dopaminah.ui.navigation.PermissionState
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {
     ComposeViewport {
-        App()
+        App(
+            permissionState = PermissionState(hasUsagePermission = true)
+        )
     }
 }